+++ components/leadPages/exitConvert.js
@@ -3,9 +3,9 @@ import Button from 'react-bootstrap/Button';
 import Modal from 'react-bootstrap/Modal';
 import Row from 'react-bootstrap/Row';
 import Col from 'react-bootstrap/Col';
-import Form from 'react-bootstrap/Form';
-import { FormValidation } from 'calidation';
-import { WHATSAPP_DISPLAY_NUMBER, WHATSAPP_MESSAGE_NUMBER } from '../../constants';
+// import Form from 'react-bootstrap/Form';
+// import { FormValidation } from 'calidation';
+// import { WHATSAPP_DISPLAY_NUMBER, WHATSAPP_MESSAGE_NUMBER } from '../../constants';
 
 class ExitConvert extends React.Component {
 
@@ -49,71 +49,13 @@ class ExitConvert extends React.Component {
 
     render() {
 
-        const formConfig = {
-            full_name: {
-                isRequired: {
-                    message: 'Full name is a required field.'
-                },
-            },
-            mobile: {
-                isRequired: {
-                    message: 'Mobile is a required field.'
-                },
-                isExactLength: {
-                    message: 'Mobile number must have exactly 10 digits.',
-                    length: 10,
-                },
-                isRegexMatch: {
-                    message: 'Mobile numbers can only contain digits.',
-                    regex: /^\d{10}$/,
-                }
-            },
-            email: {
-                isRequired: {
-                    message: 'Email is a required field.'
-                },
-                isEmail: { message: 'Please enter a valid e-mail address' },
-            }
-        };
-
-        const onSubmit = async ({ fields, isValid }) => {
-            if (isValid) {
-                this.setState({ 'buttonText': 'Processing...' });
-                const response = await fetch('/api/v1/leads/', {
-                    'body': JSON.stringify({
-                        "full_name": fields.full_name,
-                        "email": fields.email,
-                        "mobile": fields.mobile,
-                        "product": document.getElementsByName("product")[0].value
-                    }),
-                    'method': 'POST',
-                    'headers': {
-                        "cache-control": "no-cache",
-                        'Accept': 'application/json',
-                        'Content-Type': 'application/json'
-                    }
-                })
-
-                if (response.ok) {
-                    gtag_report_conversion("LEAD_FORM", false);
-                    window.setTimeout(() => {
-                        window.open(`https://api.whatsapp.com/send?phone=${WHATSAPP_MESSAGE_NUMBER}&text=Hi%21%20I%20just%20submitted%20a%20request%20form%20on%20your%20website.%0d%0aI%20am%20interested%20in%20your%20product.&source=&data=`)
-                    }, 1000)
-                } else {
-                    response.text().then(result => { console.log(result) });
-                }
-
-                this.setState({
-                    'convertSuccess': 'x',
-                })
-            }
-        };
+     
 
         return (
             <>
                 <Modal
                     {...this.props}
-                    size="lg"
+                    size="md"
                     aria-labelledby="contained-modal-title-vcenter"
                     dialogClassName="pop-convert--modal"
                     backdrop="static"
@@ -124,14 +66,19 @@ class ExitConvert extends React.Component {
                     onHide={() => { this.setState({ 'showExitConvert': false }) }}
                 >
                     <Modal.Body>
-                        <Row style={{ marginTop: '50px' }}></Row>
+                     {/* <Row style={{ marginTop: '50px' }}></Row> */}
+                        <div className="modal_header">
+                                <Button type="button" className="close" onClick={(e) => { this._handleClose(e) }} aria-label="Close">
+                                <span aria-hidden="true">×</span>
+                                </Button>
+                        </div>
                         <Row className="d-flex justify-content-center align-items-center">
-                            <Col md={{ 'span': '5' }} className="d-flex justify-content-center flex-column align-items-start">
+                            <Col md={{ 'span': '11' }} className="d-flex justify-content-center flex-column align-items-start">
+
 
                                 <a href="" onClick={(e) => { this._handleClose(e) }}>
                                     <img src="/static/images/logos/scfo/no-cap.png" className="pop-convert--modal-logo"></img>
                                 </a>
-
                                 <br></br><br></br>
                                 <p className="pop-convert--modal-title color__dark font__20">
                                     Not sure yet? Talk to us.
@@ -140,59 +87,17 @@ class ExitConvert extends React.Component {
                                     Get a free consultation with us. It's costs you nothing, and we can answer all your questions for you. <br></br>
                                     Our advisors are available 12 hours a day, 6 days a week.
                                 </span>
-                                <br></br><br></br>
+                                <br></br>
                                 {
-                                    this.state.convertSuccess ? <>
-                                        <p>
-                                            <Row className="text-center">
-                                                <Col>
-                                                    <h4 className="color__dark font__15">WhatsApp us directly at</h4>
-                                                    <span className="color__scfo font__30 mobile_font__15"><i className="fab fa-whatsapp"></i>&nbsp; {WHATSAPP_DISPLAY_NUMBER}</span>
-                                                    <br></br>
-                                                </Col>
-                                            </Row>
-                                            <br></br>
-                                            We've received your message. We can't wait to get started!
-                                            We will reach out in the next 30 - 90 minutes. You can call us if you need a faster response.
-                                            <br></br><br></br>
-                                            <span className="color__grey">
-                                                WhatsApp us at <a href={`https://api.whatsapp.com/send?phone=${WHATSAPP_MESSAGE_NUMBER}&text=Hi%21%20I%20just%20submitted%20a%20request%20form%20on%20your%20website.%0d%0aI%20am%20interested%20in%20your%20product.&source=&data=`} target="_blank" rel="noreferrer"><i className="fab fa-whatsapp color__whatsapp"></i> {WHATSAPP_DISPLAY_NUMBER}</a>
-                                                <br></br>
-                                                Email us at info@starterscfo.in
-                                                <br></br>
-                                                Chat with us, look for chat bubble in the bottom right corner of the screen.
-                                            </span>
-                                        </p>
-                                        <Button variant="primary" className="btn__primary btn-sm" onClick={(e) => { this._handleClose(e) }}>Continue browing</Button>
-                                    </> : <>
-                                        <FormValidation className="pop-convert--form" config={formConfig} onSubmit={onSubmit}>
-                                            {({ errors, fields, submitted }) => (
-                                                <>
-                                                    <Fragment>
-                                                        <Form.Group controlId="formBasicName">
-                                                            <Form.Control type="text" placeholder="Full Name" required name="full_name" value={fields.full_name} onChange={() => { }} />
-                                                            {submitted && errors.full_name && <small className="text-danger">{errors.full_name}</small>}
-                                                        </Form.Group>
-
-                                                        <Form.Group controlId="formBasicEmail">
-                                                            <Form.Control type="email" placeholder="Email *" required name="email" value={fields.email} onChange={() => { }} />
-                                                            {submitted && errors.email && <small className="text-danger">{errors.email}</small>}
-                                                        </Form.Group>
-
-                                                        <Form.Group controlId="formBasicMobile">
-                                                            <Form.Control type="tel" placeholder="Mobile *" required name="mobile" value={fields.mobile} onChange={() => { }} />
-                                                            {submitted && errors.mobile && <small className="text-danger">{errors.mobile}</small>}
-                                                        </Form.Group>
-                                                        <input name="product" value={this.props.product} onChange={() => { }} style={{ 'display': 'none' }}></input>
-                                                    </Fragment>
-
-                                                    <Button variant="primary" type="submit">
-                                                        {this.state.buttonText}
-                                                    </Button>
-                                                </>
-                                            )}
-                                        </FormValidation>
-                                        <br></br>
+                                     <>
+                                     	<iframe
+                                         width="100%" 
+                                        height="300" 
+                                        src="https://crm.starterscfo.com/forms/wtl/b56e9ff27f8b8b45c8be967db7edc920" 
+                                        frameBorder="0" 
+                                        sandbox="allow-top-navigation allow-forms allow-same-origin allow-popups" 
+                                        allowFullScreen>
+                                        </iframe>
                                         <p className="color__light-grey font__10">
                                             By creating an account with Starter's CFO, you agree to our <a href="/terms-of-service">terms</a> and <a href="/privacy-policy">privacy policy</a>.
                                         </p>
